workaround conditional breakpoint failure immediately after eclipse restart
Workspace project lookup initialization is a long-running job but debug
ISourceLookupParticipant does not provide a way to know if it is being
called from a background job or from UI thread. To be safe, the code
always forked initialization to a new background job. Breakpoint condition
evaluation was failing while initialization was running in background.

To workaround, implemented reflection-based hack to locate "thread context"
Job instance and corresponding IProgressMonitor. Change lookup
initialization to run on the same thread if such context Job was
available. Still fork initialization in background if no context
Job. Not exactly pretty, but appears to work.

Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>